+++ src/index.ts
@@ -3,8 +3,8 @@ export * from './components/BodyPortalSlotsContext';
 export * from './components/Button';
 export * from './components/ButtonBar';
 export * from './components/ButtonNav';
-export * from './components/Checkbox';
-export * from './components/CheckboxRAC';
+export * from './components/Checkbox/Checkbox';
+export * from './components/Checkbox/CheckboxRAC';
 export * from './components/DropdownMenu';
 export * from './components/Error';
 export * from './components/ErrorBoundary';
